refactor dynamic variable scripts to shared utils
