+++ src/app/shared/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./mock.utils";
 export * from "./utils";
+export * from "./variables.utils";
